Add next guess, iterate computer guess, fix bug
Computer's guesses now iterate properly over the next possible
guesses found using the function choose_value_from_min_max_array.

The bug was in the remove_from_array function, the comparison of response
with temp_response was flawed because it did not account for the same
responses in different positions. That was fixed by using .compact.sort

Currently, the computer accurately guesses the code in a minimum of 5 tries.