Colorize player guess, response, sleep + show turn
Now all the options, the secret code, the responses, etc. are
in proper color and format for display. In this commit, the player's
guess and the response to the player (everything in player_guess)
has been colorized and properly formatted, and a timer for sleep has also
been added for everything to be more understandable.